data_steward/requirements.txt
google-api-python-client
GoogleAppEngineCloudStorageClient==1.9.22.1
flask==0.10
flask-restful==0.3.5
Flask-FlatPages
mock>=1.0.1
cachetools>=1.1.5